requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Selenium & Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1
gspread==5.12.0
oauth2client==4.1.3
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Configuration & Security
python-dotenv>=1.0.0
# @SECURITY A02 - 안전한 비밀번호 해싱
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Pydantic for validation
pydantic>=2.5.0

# Testing
websockets>=12.0
